move-status-indicator-to-new-group
Move the status indicator to a separate group. It was in a button group, but according to the ev3 manual it is not a button, so I created a separate group for it and output the code to a separate file. It's logical.
